+++ client/src/components/Budget/CostsPanel.helpers.ts
@@ -51,3 +51,74 @@ export function rebalancePayers(
   free.forEach((id, i) => { next[id] = shares[i] ? shares[i].toFixed(2) : '' })
   return next
 }
+
+/**
+ * Split `total` equally across `members`, in whole cents.
+ *
+ * The remainder cent rotates with the item id rather than always landing on the
+ * first member, so across several expenses the rounding evens out instead of
+ * always favouring the same person.
+ */
+export function splitEqualShares(total: number, members: { user_id: number }[], itemId: number): Record<number, number> {
+  const n = members.length
+  if (n === 0) return {}
+
+  const totalCents = Math.round(total * 100)
+  const baseCents = Math.floor(totalCents / n)
+  const remainder = totalCents % n
+
+  const shares: Record<number, number> = {}
+  const sortedMembers = [...members].sort((a, b) => a.user_id - b.user_id)
+  const startIndex = itemId % n
+
+  for (let i = 0; i < n; i++) {
+    const member = sortedMembers[i]
+    const hasExtraCent = ((i - startIndex + n) % n) < remainder
+    shares[member.user_id] = (baseCents + (hasExtraCent ? 1 : 0)) / 100
+  }
+
+  return shares
+}
+
+/** One line of a receipt: what it cost and who is in on it. */
+export interface TicketItem {
+  id: string
+  name: string
+  price: string
+  participants: Set<number>
+}
+
+/** Per-person totals for a receipt split line by line, plus the receipt total. */
+export function calculateTicketShares(items: TicketItem[]): { shares: Record<number, number>; total: number } {
+  const shares: Record<number, number> = {}
+  let totalCents = 0
+
+  for (const item of items) {
+    const priceNum = parseFloat(item.price) || 0
+    const priceCents = Math.round(priceNum * 100)
+    totalCents += priceCents
+
+    const partIds = [...item.participants]
+    const n = partIds.length
+    if (n === 0) continue
+
+    const baseCents = Math.floor(priceCents / n)
+    const remainder = priceCents % n
+
+    const sortedPartIds = [...partIds].sort((a, b) => a - b)
+
+    for (let i = 0; i < n; i++) {
+      const id = sortedPartIds[i]
+      const hasExtraCent = i < remainder
+      const shareCents = baseCents + (hasExtraCent ? 1 : 0)
+      shares[id] = (shares[id] || 0) + shareCents
+    }
+  }
+
+  const finalShares: Record<number, number> = {}
+  for (const id of Object.keys(shares)) {
+    finalShares[Number(id)] = shares[Number(id)] / 100
+  }
+
+  return { shares: finalShares, total: totalCents / 100 }
+}
